apps/ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
lightgbm==4.5.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
psycopg2-binary==2.9.10
joblib==1.4.2
